requirements.txt
Flask==2.2.1
line-bot-sdk==1.8.0